requirements.txt
unittest
pytest
nose
parameterized
pytest-playwright
alembic
SQLAlchemy
